python_analyzer/requirements.txt
nltk
scikit-learn
numpy
python-dotenv
requests
google-genai